feat: Remove option from local
This change was made to make repo use the default local path for local
operations. This means that it is not dependent on the user defining a
env variable. There will always be a default local path so it does
not need to be an option anymore

This also means that the root workspace path in the config has to be an
option as we dont want to default init the local path to the default
and override a global path that is defined by the user