feat: Add 'tag' subcommand that contains add cmd
This change required me to change how I was defining my application with
the define_app! macro. I needed the application's ArgMatches to pass
though to the other subcommand. This required me to change the run
command of CliCommand.

Implemented the base function on adding a tag into the local or global
cache