feat: Add Abbrev url to add command for repos
This command required to add a new AbbrevUrl structure to represent a
'user/path' query. This was then added as a query result.

Things that have to do with a query have been moved into a query module.

Config now write a RawConfigData object instead of ConfigData as that is
what is deserialized when reading the cache.