fix: Global and Local config now created to be empty
There was a bug where the local config data was being set to default
if the config file did not exist. The default config has all the
default values for the config if no value is set. This means that the
local config would always have the default values and would overide
the global configuration that was set by the user