feat: Execute after clone hook
This change required a refactor on how commands were executed. This
changed the foreach command to conform to new refactor.

This change also shawn a light on the requirement of having a shell
configuration option so that we can execute commands against a shell.
This is required because executing a command like 'pwd ; ls', the
Command would not understand how to understand that and was failing. The
shell configuration option was added